refactor: Test Config 클래스 이름 적절히 변경
